feat: auth w identity, docker-compose, timer logs, db migrations
[Backend Scaffold, Config]
Set up models for the backend - User, Log.
Set up a DB context (EF core) and connected the backend to DB.

[Docker]
Set up a docker compose file for PgSql.

[Backend]
Utilize .Net Identity to setup the Auth for the backend (vanilla).
Write to DB the timer stats of logged in Users.